extract_read_files
