+++ cg/meta/workflow/rnafusion.py
@@ -102,12 +102,8 @@ class RnafusionAnalysisAPI(NfAnalysisAPI):
 
         for link in case_obj.links:
             sample_metadata: List[str] = self.gather_file_metadata_for_sample(link.sample)
-            fastq_r1: List[str] = NextflowHandler.extract_read_files(
-                metadata=sample_metadata, forward=True
-            )
-            fastq_r2: List[str] = NextflowHandler.extract_read_files(
-                metadata=sample_metadata, reverse=True
-            )
+            fastq_r1: List[str] = self.extract_read_files(metadata=sample_metadata, forward=True)
+            fastq_r2: List[str] = self.extract_read_files(metadata=sample_metadata, reverse=True)
             samplesheet_content: Dict[str, List[str]] = self.build_samplesheet_content(
                 case_id, fastq_r1, fastq_r2, strandedness
             )
